+++ internal/tofu/context_import.go
@@ -108,6 +108,54 @@ func NewImportResolver() *ImportResolver {
 	return &ImportResolver{imports: make(map[string]EvaluatedConfigImportTarget)}
 }
 
+// ValidateImportIDs is used during the validation phase to validate the import IDs of all import targets.
+// This function works similarly to ExpandAndResolveImport, but it only validates the IDs of the import targets and does not modify the EvalContext.
+// We only validate the IDs during the validation phase. Otherwise, we might cause a false positive,
+// since we do not know if the user intends to use the '-generate-config-out' option to generate additional configuration, which would make invalid Addresses valid
+func (ri *ImportResolver) ValidateImportIDs(ctx context.Context, importTarget *ImportTarget, evalCtx EvalContext) tfdiags.Diagnostics {
+	var diags tfdiags.Diagnostics
+
+	// The import block expressions are declared within the root module.
+	// We need to explicitly use the context with the path of the root module, so that all references will be
+	// relative to the root module
+	rootCtx := evalCtx.WithPath(addrs.RootModuleInstance)
+
+	if importTarget.Config.ForEach != nil {
+		const unknownsNotAllowed = false
+		const tupleAllowed = true
+
+		// The import target has a for_each attribute, so we need to expand it
+		forEachVal, evalDiags := evaluateForEachExpressionValue(ctx, importTarget.Config.ForEach, rootCtx, unknownsNotAllowed, tupleAllowed, nil)
+		diags = diags.Append(evalDiags)
+		if diags.HasErrors() {
+			return diags
+		}
+
+		// We are building an instances.RepetitionData based on each for_each key and val combination
+		var repetitions []instances.RepetitionData
+
+		it := forEachVal.ElementIterator()
+		for it.Next() {
+			k, v := it.Element()
+			repetitions = append(repetitions, instances.RepetitionData{
+				EachKey:   k,
+				EachValue: v,
+			})
+		}
+
+		for _, keyData := range repetitions {
+			_, evalDiags = evaluateImportIdExpression(importTarget.Config.ID, evalCtx, keyData)
+			diags = diags.Append(evalDiags)
+		}
+	} else {
+		// The import target is singular, no need to expand
+		_, evalDiags := evaluateImportIdExpression(importTarget.Config.ID, evalCtx, EvalDataForNoInstanceKey)
+		diags = diags.Append(evalDiags)
+	}
+
+	return diags
+}
+
 // ExpandAndResolveImport is responsible for two operations:
 // 1. Expands the ImportTarget (originating from an import block) if it contains a 'for_each' attribute.
 // 2. Goes over the expanded imports and resolves the ID and address, when we have the context necessary to resolve